requirements.txt
playwright
python-dotenv
apscheduler
confluent-kafka
vaderSentiment
transformers
torch
psycopg2-binary

#playwright install chromium